+++ app/src/main/java/com/example/selvamoneymanager/trans/TransactionRow.kt
@@ -1,16 +1,16 @@
 package com.example.selvamoneymanager.trans
 
 data class TransactionRow(
-    val id: Int,
+    val id: Long,
     val type: String,
     val dateMillis: Long,
     val amount: Double,
     val category: String?,
     val description: String,
-    val accountId: Int?,
+    val accountId: Long?,
     val accountName: String?,        // for income/expense
-    val fromAccountId: Int?,         // for transfer
+    val fromAccountId: Long?,         // for transfer
     val fromAccountName: String?,    // for transfer
-    val toAccountId: Int?,           // for transfer
+    val toAccountId: Long?,           // for transfer
     val toAccountName: String?       // for transfer
 )
